+++ plugins/build.gradle
@@ -34,6 +34,7 @@ dependencies {
     implementation fileTree(dir: 'libs', include: '*.jar')
     testImplementation project(":framework")
     testImplementation project(":framework").sourceSets.test.output
+    implementation project(":crypto")
     implementation group: 'info.picocli', name: 'picocli', version: '4.6.3'
     implementation group: 'com.typesafe', name: 'config', version: '1.3.2'
     implementation group: 'me.tongfei', name: 'progressbar', version: '0.9.3'
